Activity Log - creating a new component for activity description so that it can be re-used for standard activities and aggregated activities.
